# docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Phlag API",
        "description": "API for managing feature flags, environments, and evaluations within Phlag.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/v1/projects/{project}/flags": {
            "get": {
                "tags": [
                    "Flags"
                ],
                "summary": "List flags for a project",
                "description": "List flags for the given project.",
                "operationId": "listProjectFlags",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "description": "Project key.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number to retrieve.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of results per page (1-100).",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of flags.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FlagCollection"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication is required.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Project not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Too many requests.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Unexpected server error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Flags"
                ],
                "summary": "Create a flag for a project",
                "description": "Store a newly created flag.",
                "operationId": "createProjectFlag",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "description": "Project key.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FlagCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Flag created.",
                        "headers": {
                            "Location": {
                                "description": "URI of the created flag.",
                                "schema": {
                                    "type": "string",
                                    "format": "uri"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FlagResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Request body contained malformed JSON.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication is required.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Project not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Unexpected server error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PaginationLinks": {
                "description": "Pagination links keyed by relation name.",
                "type": "object",
                "additionalProperties": {
                    "type": "string",
                    "nullable": true
                }
            },
            "PaginationMeta": {
                "description": "Pagination metadata returned alongside resource collections.",
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "minimum": 1
                    },
                    "per_page": {
                        "type": "integer",
                        "minimum": 1
                    },
                    "total": {
                        "type": "integer",
                        "minimum": 0
                    },
                    "last_page": {
                        "type": "integer",
                        "minimum": 1,
                        "nullable": true
                    }
                },
                "type": "object",
                "additionalProperties": {}
            },
...
            "FlagVariant": {
                "description": "Flag variant configuration.",
                "required": [
                    "key"
                ],
                "properties": {
                    "key": {
                        "type": "string",
                        "example": "variant-a"
                    },
                    "weight": {
                        "type": "integer",
                        "maximum": 100,
                        "minimum": 0,
                        "nullable": true
                    },
                    "payload": {
                        "type": "object",
                        "nullable": true,
                        "additionalProperties": {}
                    }
                },
                "type": "object"
            },
            "FlagRule": {
                "description": "Flag evaluation rule definition.",
                "required": [
                    "match",
                    "variant"
                ],
                "properties": {
                    "match": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "variant": {
                        "type": "string"
                    },
                    "rollout": {
                        "type": "integer",
                        "maximum": 100,
                        "minimum": 0,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Flag": {
                "description": "Flag representation.",
                "required": [
                    "id",
                    "project_id",
                    "key",
                    "name",
                    "is_enabled",
                    "variants",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "project_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "key": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "is_enabled": {
                        "type": "boolean"
                    },
                    "variants": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/FlagVariant"
                        }
                    },
                    "rules": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/FlagRule"
                        },
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "FlagResponse": {
                "description": "Single flag response envelope.",
                "required": [
                    "data"
                ],
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/Flag"
                    }
                },
                "type": "object"
            },
            "FlagCollection": {
                "description": "Paginated flag collection response.",
                "required": [
                    "data",
                    "links",
                    "meta"
                ],
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Flag"
                        }
                    },
                    "links": {
                        "$ref": "#/components/schemas/PaginationLinks"
                    },
                    "meta": {
                        "$ref": "#/components/schemas/PaginationMeta"
                    }
                },
                "type": "object"
            },
...
            "FlagCreateRequest": {
                "description": "Payload for creating a flag.",
                "required": [
                    "key",
                    "name",
                    "variants"
                ],
                "properties": {
                    "key": {
                        "type": "string",
                        "maxLength": 64,
                        "pattern": "^[a-z0-9][a-z0-9-]*$"
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "is_enabled": {
                        "type": "boolean",
                        "nullable": true
                    },
                    "variants": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/FlagVariant"
                        },
                        "minItems": 1
                    },
                    "rules": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/FlagRule"
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "ErrorViolation": {
                "description": "Details of a single validation violation.",
                "required": [
                    "field",
                    "message"
                ],
                "properties": {
                    "field": {
                        "description": "Field path that failed validation.",
                        "type": "string"
                    },
                    "message": {
                        "description": "Human-readable description of the validation error.",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ErrorEnvelope": {
                "description": "Standardized error payload returned by the Phlag API.",
                "required": [
                    "code",
                    "message",
                    "status"
                ],
                "properties": {
                    "code": {
                        "description": "Machine-readable error identifier.",
                        "type": "string",
                        "enum": [
                            "bad_request",
                            "conflict",
                            "forbidden",
                            "http_error",
                            "method_not_allowed",
                            "not_implemented",
                            "rate_limited",
                            "resource_not_found",
                            "server_error",
                            "unauthorized",
                            "validation_failed"
                        ]
                    },
                    "message": {
                        "description": "Concise, human-readable explanation of the error.",
                        "type": "string"
                    },
                    "status": {
                        "description": "HTTP status code associated with the error.",
                        "type": "integer",
                        "format": "int32"
                    },
                    "detail": {
                        "description": "Optional detail for debugging when debug mode is enabled.",
                        "type": "string",
                        "nullable": true
                    },
                    "context": {
                        "description": "Optional key-value pairs providing additional context for the error.",
                        "type": "object",
                        "nullable": true
                    },
                    "violations": {
                        "description": "List of validation violations when the error represents validation failures.",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ErrorViolation"
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "description": "Envelope for all error responses returned by the Phlag API.",
                "required": [
                    "error"
                ],
                "properties": {
                    "error": {
                        "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "description": "JWT issued via POST /v1/auth/token.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "BearerAuth": []
        }
    ],
...
}